+++ spec/test_app/db/migrate/20111230203456_created_at_specific_dates.rb
@@ -1,0 +1,11 @@
+class CreatedAtSpecificDates < ActiveRecord::Migration
+  def up
+    NobelPrizeWinner.all.each do |npw|
+      npw.created_at = DateTime.new(2011, 1, npw.id, 11, npw.id, 12)
+      npw.save
+    end
+  end
+
+  def down
+  end
+end